requirements.txt
streamlit
pandas
plotly
ollama
python-dateutil